+++ service/controller/app/resource/chart/resource.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
-	"github.com/giantswarm/app/v3/pkg/annotation"
+	"github.com/giantswarm/app/v4/pkg/annotation"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"